+++ verik-importer/src/main/kotlin/io/verik/importer/serialize/source/SourceSerializerVisitor.kt
@@ -17,6 +17,7 @@
 package io.verik.importer.serialize.source
 
 import io.verik.importer.ast.element.common.EElement
+import io.verik.importer.ast.element.declaration.ECompanionObject
 import io.verik.importer.ast.element.declaration.EEnum
 import io.verik.importer.ast.element.declaration.EEnumEntry
 import io.verik.importer.ast.element.declaration.EKtClass
@@ -39,6 +40,10 @@ class SourceSerializerVisitor(
 
     override fun visitKtClass(`class`: EKtClass) {
         DeclarationSerializer.serializeClass(`class`, serializeContext)
+    }
+
+    override fun visitCompanionObject(companionObject: ECompanionObject) {
+        DeclarationSerializer.serializeCompanionObject(companionObject, serializeContext)
     }
 
     override fun visitEnum(enum: EEnum) {